requirements.txt
fastapi
pandas
Pillow
scikit_learn
uvicorn
timm
torchvision
python-multipart

